[feat] Delete & Put 기능 구현

+++ api/src/main/java/org/coteis/repository/CoteisRepository.java
@@ -3,5 +3,5 @@ package org.coteis.repository;
 import org.coteis.domain.Article;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CoteisRepository extends JpaRepository<Article, Integer> {
+public interface CoteisRepository extends JpaRepository<Article, Long> {
 }
